Made the logic that assigns createc blueprints to user groups extensible via ICreatedBluePrintAssignmentToUserGroupBehaviour. Migrated existing logic to the shipping implementation DefaultCreatedBluePrintAssignmentToUserGroupBehaviour.
