+++ src/Umbraco.Core/Security/ICreatedBluePrintAssignmentToUserGroupBehaviour.cs
@@ -1,0 +1,10 @@
+﻿using Umbraco.Core.Models;
+using Umbraco.Core.Models.Membership;
+
+namespace Umbraco.Core.Security
+{
+    public interface ICreatedBluePrintAssignmentToUserGroupBehaviour
+    {
+        void AssignUserGroupsToBlueprint(IContent blueprint, IUser currentUser);
+    }
+}
