+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -237,6 +237,8 @@
     <Compile Include="PropertyEditors\DateTimeConfiguration.cs" />
     <Compile Include="Migrations\Upgrade\V_8_0_0\RenameLabelAndRichTextPropertyEditorAliases.cs" />
     <Compile Include="PublishedModelFactoryExtensions.cs" />
+    <Compile Include="Security\DefaultCreatedBluePrintAssignmentToUserGroupBehaviour.cs" />
+    <Compile Include="Security\ICreatedBluePrintAssignmentToUserGroupBehaviour.cs" />
     <Compile Include="Services\PropertyValidationService.cs" />
     <Compile Include="Composing\TypeCollectionBuilderBase.cs" />
     <Compile Include="TypeLoaderExtensions.cs" />
